requirements.txt
flask
pymysql
python-dotenv
